+++ examples/data-sources/sse_application_categories/data-source.tf
@@ -1,0 +1,5 @@
+data "sse_application_categories" "all" {}
+
+output "all_application_categories" {
+  value = data.sse_application_categories.all.application_categories
+}