+++ examples/data-sources/sse_applications/data-source.tf
@@ -1,0 +1,6 @@
+data "sse_applications" "all" {}
+
+output "all_applications" {
+  value = data.sse_applications.all.applications
+}
+